+++ packages/settings-component/src/stories/input-fields/array/array-input.ts
@@ -22,6 +22,9 @@ export class ArrayInput extends LitElement {
 	schema: any = {}
 
 	@property()
+	required?: boolean = false
+
+	@property()
 	handleInlangProjectChange: (value: Array<string>, key: string, moduleId?: string) => void =
 		() => {}
 
@@ -35,12 +38,13 @@ export class ArrayInput extends LitElement {
 		) {
 			return html`
 				<language-tags-input
-					exportparts="property, property-title, property-paragraph"
+					exportparts="property, property-title, property-paragraph, button"
 					.property=${this.property}
 					.moduleId=${this.moduleId}
 					.value=${this.value}
 					.schema=${this.schema}
 					.handleInlangProjectChange=${this.handleInlangProjectChange}
+					.required=${this.required}
 				></language-tags-input>
 			`
 		} else if (this.property === "variableReferencePattern") {
@@ -52,17 +56,19 @@ export class ArrayInput extends LitElement {
 					.value=${this.value}
 					.schema=${this.schema}
 					.handleInlangProjectChange=${this.handleInlangProjectChange}
+					.required=${this.required}
 				></reference-pattern-input>
 			`
 		} else {
 			return html`
 				<default-array-input
-					exportparts="property, property-title, property-paragraph"
+					exportparts="property, property-title, property-paragraph, button"
 					.property=${this.property}
 					.moduleId=${this.moduleId}
 					.value=${this.value}
 					.schema=${this.schema}
 					.handleInlangProjectChange=${this.handleInlangProjectChange}
+					.required=${this.required}
 				></default-array-input>
 			`
 		}