+++ packages/settings-component/src/stories/input-fields/string/string-input.ts
@@ -1,23 +1,31 @@
 import { css, html, LitElement } from "lit"
 import { customElement, property } from "lit/decorators.js"
-//import { baseStyling } from "../../../styling/base.js"
+import "./../../field-header.js"
 
 @customElement("string-input")
 export class StringInput extends LitElement {
 	static override styles = [
-		//baseStyling,
 		css`
+			.property {
+				display: flex;
+				flex-direction: column;
+				gap: 12px;
+			}
+			h3 {
+				margin: 0;
+				font-size: 14px;
+				font-weight: 800;
+			}
 			.help-text {
-				font-size: 0.8rem;
+				font-size: 14px;
 				color: var(--sl-input-help-text-color);
 				margin: 0;
+				line-height: 1.5;
 			}
 			.description-container {
 				display: flex;
 				flex-direction: column;
 				gap: 4px;
-				margin-bottom: 1rem;
-				margin-top: 0.8rem;
 			}
 		`,
 	]
@@ -35,6 +43,9 @@ export class StringInput extends LitElement {
 	schema: any = {}
 
 	@property()
+	required?: boolean = false
+
+	@property()
 	handleInlangProjectChange: (value: string, key: string, moduleId?: string) => void = () => {}
 
 	private get _description(): string | undefined {
@@ -42,18 +53,22 @@ export class StringInput extends LitElement {
 	}
 
 	private get _examples(): string | undefined {
-		return this.schema.examples ? "Example: " + JSON.stringify(this.schema.examples) : undefined
+		return this.schema.examples
+	}
+
+	private get _title(): string | undefined {
+		return this.schema.title || undefined
 	}
 
 	override render() {
-		return html` <div part="property">
-			<h3 part="property-title">${this.property}</h3>
-			<div class="description-container">
-				${this._description &&
-				html`<p part="property-paragraph" class="help-text">${this._description}</p>`}
-				${this._examples &&
-				html`<p part="property-paragraph" class="help-text">${this._examples}</p>`}
-			</div>
+		return html` <div part="property" class="property">
+			<field-header
+				.fieldTitle=${this._title ? this._title : this.property}
+				.description=${this._description}
+				.examples=${this._examples}
+				.optional=${this.required ? false : true}
+				exportparts="property-title, property-paragraph"
+			></field-header>
 			<sl-input
 				value=${this.value}
 				size="small"