+++ packages/github-lint-action/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @inlang/github-lint-action
 
+## 0.2.18
+
+### Patch Changes
+
+- Updated dependencies [c068dd2]
+  - @inlang/sdk@0.31.0
+
+## 0.2.17
+
+### Patch Changes
+
+- Updated dependencies [9b26a31]
+  - @inlang/sdk@0.30.0
+
 ## 0.2.16
 
 ### Patch Changes