+++ packages/plugins/i18next/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @inlang/plugin-i18next
 
+## 4.13.24
+
+### Patch Changes
+
+- Updated dependencies [c068dd2]
+  - @inlang/sdk@0.31.0
+
+## 4.13.23
+
+### Patch Changes
+
+- Updated dependencies [9b26a31]
+  - @inlang/sdk@0.30.0
+
 ## 4.13.22
 
 ### Patch Changes