+++ packages/plugins/m-function-matcher/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @inlang/plugin-m-function-matcher
 
+## 0.9.6
+
+### Patch Changes
+
+- Updated dependencies [c068dd2]
+  - @inlang/sdk@0.31.0
+
+## 0.9.5
+
+### Patch Changes
+
+- Updated dependencies [9b26a31]
+  - @inlang/sdk@0.30.0
+
 ## 0.9.4
 
 ### Patch Changes