+++ packages/plugins/next-intl/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @inlang/plugin-next-intl
 
+## 1.3.15
+
+### Patch Changes
+
+- Updated dependencies [c068dd2]
+  - @inlang/sdk@0.31.0
+
+## 1.3.14
+
+### Patch Changes
+
+- Updated dependencies [9b26a31]
+  - @inlang/sdk@0.30.0
+
 ## 1.3.13
 
 ### Patch Changes