+++ packages/telemetry/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @inlang/telemetry
 
+## 0.3.21
+
+### Patch Changes
+
+- Updated dependencies [c068dd2]
+  - @inlang/sdk@0.31.0
+
+## 0.3.20
+
+### Patch Changes
+
+- Updated dependencies [9b26a31]
+  - @inlang/sdk@0.30.0
+
 ## 0.3.19
 
 ### Patch Changes